refactor: SseEmitterWrapper service 패키지로 이동
